Changed updateSettings and getSettings They are now in settingsController instead of dataController Changed the routes in server.ts Verified changes in java-server Removed unused code from ProducerMetricController
